# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings/menu-template:
    get:
      operationId: getMenuTemplate
      tags: [menu]
      summary: Get the active customer menu template
      responses:
        "200":
          description: Current menu template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuTemplateSetting"
    put:
      operationId: updateMenuTemplate
      tags: [menu]
      summary: Update the active customer menu template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuTemplateSetting"
      responses:
        "200":
          description: Menu template updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuTemplateSetting"

  # ── Dashboard Auth ───────────────────────────────────────────────────
  /dashboard/auth/login:
    post:
      operationId: dashboardLogin
      tags: [dashboard-auth]
      summary: Login to dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardUser"
        "401":
          description: Invalid credentials
  /dashboard/auth/me:
    get:
      operationId: dashboardMe
      tags: [dashboard-auth]
      summary: Get current dashboard user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardUser"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    DashboardCredentials:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    DashboardUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, employee]
      required: [id, username, role]

    # ── Orders ────────────────────────────────────────────────────────
...
    MenuTemplateSetting:
      type: object
      properties:
        menuTemplate:
          type: string
          enum: [classic, modern]
          default: classic
      required: [menuTemplate]
...